+++ src/ast/types.rs
@@ -165,7 +165,7 @@ impl From<char> for Char {
 
 impl fmt::Display for Char {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self)
+        write!(f, "{}", self.0)
     }
 }
 
@@ -239,26 +239,6 @@ impl fmt::Display for FunctionPtr {
     }
 }
 
-/// Unit struct that represents the `Self` type.
-#[derive(Debug, Clone, PartialEq, Hash, Eq, PartialOrd, Ord)]
-pub struct SelfType;
-
-impl fmt::Display for SelfType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Self")
-    }
-}
-
-/// Unit struct that represents the unit type `()`.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Unit;
-
-impl fmt::Display for Unit {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "()")
-    }
-}
-
 /// Struct that represents an inferred type.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct InferredType {
@@ -271,9 +251,29 @@ impl fmt::Display for InferredType {
     }
 }
 
+/// Unit struct that represents the `Self` type.
+#[derive(Debug, Clone, PartialEq, Hash, Eq, PartialOrd, Ord)]
+pub struct SelfType;
+
+impl fmt::Display for SelfType {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "Self")
+    }
+}
+
 /// Struct that represents the path to user-defined type (e.g., struct, enum and trait)
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct TypePath {
     pub(crate) associated_type_path_prefix_opt: Option<Vec<Identifier>>,
     pub(crate) type_name: Identifier,
+}
+
+/// Unit struct that represents the unit type `()`.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Unit;
+
+impl fmt::Display for Unit {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "()")
+    }
 }
